+++ zeebe/exporters/app-integrations-exporter/src/main/java/io/camunda/exporter/appint/transport/HttpTransportConfig.java
@@ -1,0 +1,15 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
+ * one or more contributor license agreements. See the NOTICE file distributed
+ * with this work for additional information regarding copyright ownership.
+ * Licensed under the Camunda License 1.0. You may not use this file
+ * except in compliance with the Camunda License 1.0.
+ */
+package io.camunda.exporter.appint.transport;
+
+public record HttpTransportConfig(
+    String url,
+    Authentication authentication,
+    int maxRetries,
+    long retryDelayMs,
+    long requestTimeoutMs) {}
