+++ zeebe/exporters/app-integrations-exporter/src/main/java/io/camunda/exporter/appint/transport/Transport.java
@@ -1,0 +1,15 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
+ * one or more contributor license agreements. See the NOTICE file distributed
+ * with this work for additional information regarding copyright ownership.
+ * Licensed under the Camunda License 1.0. You may not use this file
+ * except in compliance with the Camunda License 1.0.
+ */
+package io.camunda.exporter.appint.transport;
+
+import io.camunda.zeebe.util.CloseableSilently;
+import java.util.List;
+
+public interface Transport<T> extends CloseableSilently {
+  void send(List<T> events);
+}
